Use full URL in RSS feed
Hugo creates full paths from relative paths for HTML pages by default,
but not for the RSS feed. This snippet converts relative paths to full
paths for the RSS feed. The relative paths in the RSS feed meant the
links didn't work in RSS readers and in the emails that are built from
them.